Move Ast types to their own file
