+++ src/ast.ts
@@ -1,0 +1,13 @@
+export type Line = Array<string | Inline>;
+
+export interface Block {
+	tag: string;
+	head: Line;
+	children: Block[];
+}
+
+export interface Inline {
+	tag: string;
+	arguments: Line[];
+	closed: boolean;
+}
